requirements.txt
aiohttp~=3.9.0
Brotli~=1.1.0
yarl~=1.9.3
pyttsx3~=2.90
azure-cognitiveservices-speech==1.34.1